requirements-django.txt
django==4.0.*
django-bootstrap-v5==1.0.*
django-environ==0.8.*
psycopg2-binary==2.9.*
whitenoise==5.3.*
honcho==1.1.*
gunicorn==20.1.*
